#294 Move enumerate code to SCNNode extension.
