+++ ios/RNMagicScript/components/Utils/Extensions/SCNNode+Enumerate.swift
@@ -1,0 +1,40 @@
+//
+//  Copyright (c) 2019 Magic Leap, Inc. All Rights Reserved
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+// 
+
+import Foundation
+import SceneKit
+
+extension SCNNode {
+    @objc func enumerateComponentsHierarchy(_ block: (TransformNode) -> Void) {
+        childNodes.forEach { (node) in
+            node.enumerateComponentsHierarchy(block)
+        }
+
+        if let transformNode = self as? TransformNode {
+            block(transformNode)
+        }
+    }
+
+    @objc func enumeratePostOrder(block: @escaping (TransformNode) -> Void) {
+        childNodes.forEach { child in
+            child.enumeratePostOrder(block: block)
+        }
+
+        if let transformNode = self as? TransformNode {
+            block(transformNode)
+        }
+    }
+}
